🔨 Refactor : 타입정리 & 초기값 정리
