+++ src/Component/Mission/MissionRule.tsx
@@ -1,7 +1,5 @@
 import { styled } from 'styled-components';
-import { PageSort } from '../../Type/MissionType';
-import { MindsType } from '../../Type/Mind';
-import { getMindInfoType } from '../../API/Mind';
+import { MindsType, PageSort } from '../../Type/Mind';
 interface MissionRuleType {
   getMindInfoData: MindsType;
   passsort: PageSort;
