+++ src/Page/GroupPage/GroupPageBarrel.ts
@@ -4,7 +4,6 @@ import { useNavigate } from 'react-router-dom';
 
 import useLoginCheck from '../../Hooks/useLoginCheck';
 import { GroupHeader } from '../../Component/Mission/GroupHeader';
-import LikeBind from '../../Type/LikeBind';
 import Comment from './Comment';
 import DivideBaS from '../../Component/Mission/DivideBa';
 import GroupPost from './GroupPost';
@@ -19,7 +18,6 @@ export {
   useNavigate,
   useLoginCheck,
   GroupHeader,
-  type LikeBind,
   Comment,
   DivideBaS,
   GroupPost,
