+++ src/Page/MyPage/MypageBarrel.ts
@@ -13,7 +13,6 @@ import scrollTop from '../../Hooks/scrollTop';
 import { GetUser } from '../../Type/User';
 import { getMyList } from '../../API/Mind';
 import { Mylist } from '../../Type/Mind';
-import { GroupInfoType } from '../../Type/MissionType';
 import { initUser } from '../../data/initialData';
 import { getUser, logoutUser } from '../../API/Users';
 import { MyListContextType, MyListContext } from '../../API/Context';
@@ -23,4 +22,4 @@ export { Arrow_Left_B, 기본프로필, Info_icon_B };
 export { ArticleTab, CurrentMind, EndMindList, ConfirmModal };
 export { scrollTop, getUser, getMyList, logoutUser };
 export { initUser, MyListContext };
-export type { GetUser, Mylist, GroupInfoType, MyListContextType };
+export type { GetUser, Mylist, MyListContextType };
