--- src/Type/LikeBind.ts
@@ -1,6 +1,0 @@
-type LikeBind = {
-  isLiked: boolean;
-  setIsLiked: React.Dispatch<React.SetStateAction<boolean>>;
-};
-
-export default LikeBind;
