+++ src/Type/Mind.ts
@@ -1,22 +1,22 @@
-// 그룹페이지 정보
-export interface MindPageInfo {
+export type PageSort = 'Page' | 'Intro' | 'Create';
+
+export interface CommonMind {
   mindId: number;
   mindTypeName: string;
   name: string;
   userCount: number;
+  introduce: string;
   writeFormat: string;
-  doneToday: boolean;
+}
+
+// 그룹페이지 정보
+export interface MindPageInfo extends CommonMind {
+  isDoneToday: boolean;
   count: 0 | 1 | 2 | 3;
 }
 
 // 그룹 소개 정보
-export interface MindIntroInfo {
-  mindId: number;
-  mindTypeName: string;
-  name: string;
-  introduce: string;
-  userCount: number;
-  writeFormat: string;
+export interface MindIntroInfo extends CommonMind {
   canJoin: number;
 }
 
@@ -33,7 +33,7 @@ export interface isDoneSingle {
 
 // 당일 전체 참여한 작심 인증 여부
 export interface isDone {
-  doneToday: boolean;
+  isDoneToday: boolean;
   joinedMindId: number;
 }
 
@@ -42,10 +42,10 @@ export interface Mylist {
   mindId: number;
   mindTypeName: string;
   name: string;
+  isDoneToday: boolean;
   count: 0 | 1 | 2 | 3;
   boardCount: number;
   myListImage: string;
-  isDoneToday: boolean;
 }
 
 export interface EndMindType {
@@ -55,15 +55,4 @@ export interface EndMindType {
   boardCount: number;
 }
 
-// 그룹 인트로 + 페이지 - FIXME: 고도화때 삭제
-export interface MindsType {
-  mindId: number;
-  mindTypeName: string;
-  name: string;
-  introduce: string;
-  userCount: number;
-  writeFormat?: string;
-  isDoneToday?: boolean;
-  count?: number;
-  canJoin?: number;
-}
+export type MindsType = MindPageInfo | MindIntroInfo;
