+++ src/data/initialData.ts
@@ -1,5 +1,5 @@
 import { GetUser } from '../Type/User';
-import { MindsType, Mylist } from '../Type/Mind';
+import { Mylist } from '../Type/Mind';
 
 export const initMyList: Mylist[] = [];
 
@@ -8,16 +8,4 @@ export const initUser: GetUser = {
   nickname: '',
   profileImage: `${process.env.PUBLIC_URL}/Chips_signout.png`,
   roles: '',
-};
-
-export const initMind: MindsType = {
-  mindId: 0,
-  mindTypeName: '',
-  name: '',
-  introduce: '',
-  userCount: 0,
-  writeFormat: '',
-  isDoneToday: false,
-  count: 0,
-  canJoin: 0,
 };
